Moved IVisualizationCanvas to the plugin
