+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/interfaces/ITableRowFilter.java
@@ -1,4 +1,4 @@
-package com.net2plan.interfaces;
+package com.net2plan.gui.plugins.networkDesign.interfaces;
 
 import com.google.common.collect.Sets;
 import com.net2plan.interfaces.networkDesign.*;
