+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/interfaces/ITopologyCanvasEdge.java
@@ -1,4 +1,4 @@
-package com.net2plan.interfaces;
+package com.net2plan.gui.plugins.networkDesign.interfaces;
 
 import com.net2plan.interfaces.networkDesign.Link;
