+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/interfaces/IVisualizationCallback.java
@@ -10,10 +10,11 @@
  ******************************************************************************/
 
 
-package com.net2plan.interfaces.networkDesign;
+package com.net2plan.gui.plugins.networkDesign.interfaces;
 
 import com.net2plan.interfaces.ITopologyCanvas;
-import com.net2plan.interfaces.ITopologyCanvasVertex;
+import com.net2plan.interfaces.networkDesign.NetPlan;
+import com.net2plan.interfaces.networkDesign.NetworkElement;
 import com.net2plan.internal.Constants;
 
 import java.awt.*;
