+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/topologyPane/jung/GUILink.java
@@ -13,8 +13,7 @@
 package com.net2plan.gui.plugins.networkDesign.topologyPane.jung;
 
 import com.net2plan.gui.plugins.networkDesign.visualizationControl.VisualizationConstants;
-import com.net2plan.interfaces.ITopologyCanvasEdge;
-import com.net2plan.interfaces.ITopologyCanvasVertex;
+import com.net2plan.gui.plugins.networkDesign.interfaces.ITopologyCanvasEdge;
 import com.net2plan.interfaces.networkDesign.*;
 
 import java.awt.*;
