+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/topologyPane/jung/JUNGCanvas.java
@@ -15,10 +15,10 @@ import com.net2plan.gui.plugins.networkDesign.topologyPane.TopologyPanel;
 import com.net2plan.gui.plugins.networkDesign.topologyPane.jung.osmSupport.state.OSMStateManager;
 import com.net2plan.gui.plugins.networkDesign.visualizationControl.VisualizationConstants;
 import com.net2plan.interfaces.ITopologyCanvas;
-import com.net2plan.interfaces.ITopologyCanvasEdge;
+import com.net2plan.gui.plugins.networkDesign.interfaces.ITopologyCanvasEdge;
 import com.net2plan.interfaces.ITopologyCanvasPlugin;
 import com.net2plan.gui.plugins.GUINetworkDesign;
-import com.net2plan.interfaces.ITopologyCanvasVertex;
+import com.net2plan.gui.plugins.networkDesign.interfaces.ITopologyCanvasVertex;
 import com.net2plan.interfaces.networkDesign.Configuration;
 import com.net2plan.interfaces.networkDesign.Node;
 import com.net2plan.internal.CommandLineParser;
