+++ Net2Plan-GUI/Net2Plan-GUI-Plugins/Net2Plan-NetworkDesign/src/main/java/com/net2plan/gui/plugins/networkDesign/viewEditTopolTables/tableVisualizationFilters/TBFToFromCarriedTraffic.java
@@ -1,7 +1,7 @@
 package com.net2plan.gui.plugins.networkDesign.viewEditTopolTables.tableVisualizationFilters;
 
 import com.google.common.collect.Sets;
-import com.net2plan.interfaces.ITableRowFilter;
+import com.net2plan.gui.plugins.networkDesign.interfaces.ITableRowFilter;
 import com.net2plan.interfaces.networkDesign.*;
 import com.net2plan.utils.Pair;
 import com.net2plan.utils.Triple;
